requirements.txt
streamlit
python-dotenv
SQLAlchemy
sqlalchemy-cockroachdb